feat: command property of a step replaces the exec and args properties
It turned out being troublesome to list arguments in an array of strings. Previously, the exacutable
to run was a string under the exec property of a step and the arguments were list of strings under
the args property of a step. From Dozer 3, there is a single string property command which should
include the entire command to run, including the executable and arguments. Just like it it was input
into the actual command line.

BREAKING CHANGE: exec and args properties are obsolete, which changes the pipeline config syntax