feat: the runner and UI are now separate applications communicating over websockets
This is a major refactor of Dozer towards the next version. The UI and runner have been separated
into two standalone applications which comes with a couple of benefits: First the runner itself
doesn't come with the UI code burden if not necessary. Second, UI now communicates with the runner
over websockets, which allows for multiple types (even third party) UI clients and displaying the
progress of the pipelines over the internet. Specify the --no-api command line argument to Dozer in
order to disable the websockets communication. Note: sensitive data should never be provided in
output logs as anyone knowing the IP and port of a running Dozer could connect and read the output.
Also the code of the runner has been cleaned up, separated into modules and commented for better
understanding.

BREAKING CHANGE: The `title` parameter of pipeline steps replaces `displayName`, but `displayName`
will be still accepted if `title` is not present. The `--gui` is replaced by `--webui` and will
start the Web UI in a browser tab if specified (instead of an Electron app as before). The `--gui`
parameter is now reserved for Native UI clients.